added flash messages in hostels,members

+++ controllers/adminUpload.controller.js
@@ -25,6 +25,10 @@ exports.updateUploadForm = async (req, res) => {
   try {
     const id = req.params.id;
     const upload = await AdminUploads.findById(id);
+    if (!upload) {
+      req.flash("error", "Cannot update");
+      return res.redirect("/admin/uploads");
+    }
 
     return res.render("adminUploads/edit", {
       link: "/uploads/" + req.params.id,
@@ -40,6 +44,10 @@ exports.updateUpload = async (req, res) => {
     const id = req.params.id;
 
     const upload = await AdminUploads.findById(id);
+    if (!upload) {
+      req.flash("error", "Cannot find uploaded file");
+      return res.redirect("/admin/uploads");
+    }
 
     var image_name = upload.image_name;
     var image = upload.image;
@@ -52,9 +60,14 @@ exports.updateUpload = async (req, res) => {
       image = req.file.filename;
     }
     const obj = { image_name, image };
-    await AdminUploads.findByIdAndUpdate(id, obj, {
+    const uploadFile=await AdminUploads.findByIdAndUpdate(id, obj, {
       runValidators: true,
     });
+    if (!uploadFile) {
+      req.flash("error", "Cannot upload file");
+      return res.redirect("/admin/uploads");
+    }
+    req.flash("success", "Successfully updated upload file");
 
     return res.redirect("/admin/uploads");
   } catch (error) {
@@ -69,7 +82,12 @@ exports.createUpload = async (req, res) => {
     var image;
     if (req.file) image = req.file.filename;
     const newUpload = new AdminUploads({ image_name, image });
-    await newUpload.save();
+    const upload=await newUpload.save();
+    if (!upload) {
+      req.flash("error", "Cannot add file");
+      return res.redirect("/admin/uploads");
+    }
+    req.flash("success", "Successfully uploaded file");
     return res.redirect("/admin/uploads");
   } catch (error) {
     console.log(error.message);
@@ -80,10 +98,16 @@ exports.deleteUpload = async (req, res) => {
   try {
     const id = req.params.id;
     const upload = await AdminUploads.findById(id);
+    if (!upload) {
+      req.flash("error", "Cannot find uploaded file");
+      return res.redirect("/admin/uploads");
+    }
     fs.unlinkSync(`uploads/adminUploads/${upload.image}`);
 
     console.log("successfully deleted!");
     await AdminUploads.findByIdAndRemove(id);
+    
+    req.flash("success", "Successfully deleted file");
     return res.redirect("/admin/uploads");
   } catch (err) {
     // handle the error
@@ -94,10 +118,18 @@ exports.deleteUpload = async (req, res) => {
 exports.deleteAllImages = async (req, res) => {
   try {
     const uploads = await AdminUploads.find({});
+    if (!uploads) {
+      req.flash("error", "Cannot find uploaded files");
+      return res.redirect("/admin/uploads");
+    }
+
     uploads.forEach((upload) => {
       if (upload.image) fs.unlinkSync(`uploads/adminUploads/${upload.image}`);
     });
     await AdminUploads.deleteMany({});
+    
+    req.flash("success", "Successfully deleted all files");
+
     console.log("successfully deleted all images");
 
     return res.redirect("/admin/uploads");
