+++ controllers/hostel.controller.js
@@ -9,6 +9,10 @@ exports.getAllHostels = async (req, res, next) => {
 
 exports.getHostel = async (req, res, next) => {
   const hostel = await Hostel.findOne({ name: req.params.name });
+  if (!hostel) {
+    req.flash("error", "Cannot find hostel");
+    return res.redirect("/admin/hostels");
+  }
 
   if (hostel == null) {
     return res.status(404).json({
@@ -40,12 +44,23 @@ exports.createHostel = async (req, res) => {
   var pic;
   if (req.file) pic = req.file.filename;
   const newHostel = new Hostel({ name, pic, contact1, contact2 });
-  await newHostel.save();
+
+  const hostel=await newHostel.save();
+  if (!hostel) {
+    req.flash("error", "Cannot add hostel");
+    return res.redirect("/admin/hostels");
+  }
+  req.flash("success", "Successfully added new hostel");
   return res.redirect("/admin/hostels");
 };
 
 exports.updateHostelForm = async (req, res) => {
   const hostel = await Hostel.findOne({ name: req.params.name });
+  if (!hostel) {
+    req.flash("error", "Cannot find this hostel");
+    return res.redirect("/admin/hostels");
+  }
+  
   return res.render("hostels/edit", {
     link: "/admin/hostels/" + req.params.name,
     hostel,
@@ -54,6 +69,10 @@ exports.updateHostelForm = async (req, res) => {
 
 exports.updateHostel = async (req, res) => {
   const hostel = await Hostel.findOne({ name: req.params.name });
+  if (!hostel) {
+    req.flash("error", "Cannot find hostel");
+    return res.redirect("/admin/hostels");
+  }
 
   var name = hostel.name;
   var pic = hostel.pic;
@@ -75,9 +94,14 @@ exports.updateHostel = async (req, res) => {
     contact2 = req.body.contact2;
   }
   const obj = { name, pic,contact1,contact2 };
-  await Hostel.findOneAndUpdate(req.params.name, obj, {
+  const hostell=await Hostel.findOneAndUpdate(req.params.name, obj, {
     runValidators: true,
   });
+  if (!hostell) {
+    req.flash("error", "Cannot add hostel");
+    return res.redirect("/admin/hostels");
+  }
+  req.flash("success", "Successfully updated hostel");
 
   return res.redirect("/admin/hostels");
 };
@@ -100,7 +124,7 @@ exports.createMember = async (req, res, next) => {
     contact2,
     email,
   };
-  Hostel.findOneAndUpdate(
+  const hostel=Hostel.findOneAndUpdate(
     { name: name },
     { $push: { management: member } },
     function (error) {
@@ -109,20 +133,33 @@ exports.createMember = async (req, res, next) => {
       }
     }
   );
+  if (!hostel) {
+    req.flash("error", "Cannot add member");
+    return res.redirect(`/admin/hostels/${name}`);
+  }
+  req.flash("success", "Successfully added new member");
 
-  return res.redirect(`/admin/hostels/${name}/addMember`);
+  return res.redirect(`/admin/hostels/${name}`);
 };
 
 exports.updateMemberForm = async (req, res) => {
   const name = req.params.name;
   const id = req.params.id;
   const hostel = await Hostel.findOne({ name: name });
+  if (!hostel) {
+    req.flash("error", "Cannot find hostel");
+    return res.redirect("/admin/hostels");
+  }
 
   var member = hostel.management;
   member = member.filter(function (object) {
     return object.id == id;
   });
   mb = member[0];
+  if (!mb) {
+    req.flash("error", "Cannot find member");
+    return res.redirect(`/admin/hostels/${name}`);
+  }
 
   return res.render("hostels/members/edit", {
     link: "/admin/hostels/" + req.params.name + "/" + id,
@@ -134,14 +171,22 @@ exports.updateMemberForm = async (req, res) => {
 exports.updateMember = async (req, res) => {
   const name = req.params.name;
 
-  const hostell = await Hostel.findOne({ name: name });
+  const hostel = await Hostel.findOne({ name: name });
+  if (!hostel) {
+    req.flash("error", "Cannot find hostel");
+    return res.redirect("/admin/hostels");
+  }
 
   const { Mname, position, priority, contact1,contact2, email } = req.body;
   const id = req.params.id;
-  var member = hostell.management;
+  var member = hostel.management;
   member = member.filter(function (object) {
     return object.id == id;
   });
+  if (!member) {
+    req.flash("error", "Cannot find member");
+    return res.redirect(`/admin/hostels/${name}`);
+  }
   var photo = member[0].photo;
 
   if (req.file) {
@@ -152,7 +197,7 @@ exports.updateMember = async (req, res) => {
     photo = req.file.filename;
   }
 
-  await Hostel.findOne({ name }).then((hostel) => {
+  const hostell=await Hostel.findOne({ name }).then((hostel) => {
     let management = hostel.management.id(id);
     management.Mname = Mname;
     management.position = position;
@@ -163,6 +208,11 @@ exports.updateMember = async (req, res) => {
     management.photo = photo;
     return hostel.save();
   });
+  if (!hostell) {
+    req.flash("error", "Cannot update member");
+    return res.redirect("/admin/hostels/"+ name);
+  }
+  req.flash("success", "Successfully updated member");
   return res.redirect("/admin/hostels/" + name);
 };
 
@@ -171,10 +221,18 @@ exports.deleteMember = async (req, res) => {
     const name = req.params.name;
     const id = req.params.id;
     const hostel = await Hostel.findOne({ name: name });
+    if (!hostel) {
+      req.flash("error", "Cannot find hostel");
+      return res.redirect("/admin/hostels");
+    }
     var member = hostel.management;
     member = member.filter(function (object) {
       return object.id == id;
     });
+    if (!member) {
+      req.flash("error", "Cannot find member");
+      return res.redirect("/admin/hostels"+name);
+    }
 
     if (member[0].photo) {
       fs.unlinkSync(`uploads/hostel/${member[0].photo}`);
@@ -182,6 +240,7 @@ exports.deleteMember = async (req, res) => {
     }
     hostel.management.pull({ _id: id });
     await hostel.save();
+    req.flash("success", "Successfully deleted member");
     return res.redirect("/admin/hostels/" + name);
   } catch (err) {
     // handle the error
@@ -194,12 +253,21 @@ exports.deleteHostelMembers = async (req, res) => {
   try {
     const name = req.params.name;
     const hostel = await Hostel.findOne({ name: name });
+    if (!hostel) {
+      req.flash("error", "Cannot find hostel");
+      return res.redirect("/admin/hostels");
+    }
     var members = hostel.management;
+    if (!members) {
+      req.flash("error", "Cannot find members");
+      return res.redirect("/admin/hostels/"+name);
+    }
     members.forEach((member) => {
       if (member.photo) fs.unlinkSync(`uploads/hostel/${member.photo}`);
     });
     hostel.management.splice(0, hostel.management.length);
     await hostel.save();
+    req.flash("success", "Successfully deleted all members");
     console.log("successfully deleted all members");
 
     return res.redirect("/admin/hostels/" + name);
@@ -214,6 +282,10 @@ exports.deleteHostel = async (req, res) => {
   try {
     const id = req.params.id;
     const hostel = await Hostel.findById(id);
+    if (!hostel) {
+      req.flash("error", "Cannot find hostel");
+      return res.redirect("/admin/hostels");
+    }
     fs.unlinkSync(`uploads/hostel/${hostel.pic}`);
     var members = hostel.management;
     members.forEach((member) => {
@@ -222,6 +294,8 @@ exports.deleteHostel = async (req, res) => {
 
     console.log("successfully deleted!");
     await Hostel.findByIdAndRemove(id);
+    req.flash("success", "Successfully deleted hostel");
+
     return res.redirect("/admin/hostels");
   } catch (err) {
     // handle the error
